backend/app/requirements.txt
fastapi
uvicorn
sqlmodel
mysqlclient
sqlalchemy
python-dotenv
